refactor: split large frontend components and improve performance patterns (#206)
Split two oversized component files into focused, single-responsibility
modules and add performance optimizations.

Component splits:
- AdminSettingsPage (688 → 36 lines): extract GeneralSection,
  UsersSection, OidcProvidersSection, ReviewsSection into
  src/pages/admin/
- NamespaceDetailView (634 → 65 lines): extract SettingsSection,
  MembersSection, ApiKeysSection into
  src/components/admin/namespace-detail/

Performance improvements:
- CatalogPage: remove redundant useEffect that duplicated fetchPlugins
  on namespace change; merge fetchTags into the search effect
- PluginCard + PluginListRow: wrap with React.memo to prevent
  unnecessary re-renders in mapped catalog lists
- Router: lazy-load all 5 admin section components with React.lazy()
  and Suspense fallback, reducing initial bundle for non-admin users

All 235 existing tests pass without modification.

Closes #206
