+++ plugwerk-server/plugwerk-server-frontend/src/router/index.tsx
@@ -16,7 +16,9 @@
  * You should have received a copy of the GNU Affero General Public License
  * along with Plugwerk. If not, see <https://www.gnu.org/licenses/>.
  */
+import { lazy, Suspense } from 'react'
 import { createBrowserRouter, Navigate } from 'react-router-dom'
+import { CircularProgress, Box } from '@mui/material'
 import { AppShell } from '../AppShell'
 import { AdminRoute } from '../components/auth/AdminRoute'
 import { ProtectedRoute } from '../components/auth/ProtectedRoute'
@@ -26,8 +28,7 @@ import { LoginPage } from '../pages/LoginPage'
 import { RegisterPage } from '../pages/RegisterPage'
 import { ForgotPasswordPage } from '../pages/ForgotPasswordPage'
 import { ResetPasswordPage } from '../pages/ResetPasswordPage'
-import { AdminSettingsPage, GeneralSection, UsersSection, OidcProvidersSection, ReviewsSection } from '../pages/AdminSettingsPage'
-import { NamespacesSection } from '../components/admin/NamespacesSection'
+import { AdminSettingsPage } from '../pages/AdminSettingsPage'
 import { ChangePasswordPage } from '../pages/ChangePasswordPage'
 import { ProfileSettingsPage } from '../pages/ProfileSettingsPage'
 import { Error404Page } from '../pages/errors/Error404Page'
@@ -37,6 +38,17 @@ import { Error503Page } from '../pages/errors/Error503Page'
 import { useAuthStore } from '../stores/authStore'
 import { ApiDocsPage } from '../pages/ApiDocsPage'
 import { OnboardingPage } from '../pages/OnboardingPage'
+
+// Lazy-loaded admin sections — only loaded when the user navigates to admin
+const GeneralSection = lazy(() => import('../pages/admin/GeneralSection').then((m) => ({ default: m.GeneralSection })))
+const UsersSection = lazy(() => import('../pages/admin/UsersSection').then((m) => ({ default: m.UsersSection })))
+const OidcProvidersSection = lazy(() => import('../pages/admin/OidcProvidersSection').then((m) => ({ default: m.OidcProvidersSection })))
+const ReviewsSection = lazy(() => import('../pages/admin/ReviewsSection').then((m) => ({ default: m.ReviewsSection })))
+const NamespacesSection = lazy(() => import('../components/admin/NamespacesSection').then((m) => ({ default: m.NamespacesSection })))
+
+function LazyFallback() {
+  return <Box sx={{ display: 'flex', justifyContent: 'center', py: 4 }}><CircularProgress size={24} /></Box>
+}
 
 function CatalogRedirect() {
   const namespace = useAuthStore((s) => s.namespace)
@@ -60,11 +72,11 @@ export const router = createBrowserRouter([
         element: <ProtectedRoute><AdminRoute><AdminSettingsPage /></AdminRoute></ProtectedRoute>,
         children: [
           { index: true,              element: <Navigate to="global-settings" replace /> },
-          { path: 'global-settings',  element: <GeneralSection /> },
-          { path: 'namespaces',       element: <NamespacesSection /> },
-          { path: 'users',            element: <UsersSection /> },
-          { path: 'oidc-providers',   element: <OidcProvidersSection /> },
-          { path: 'reviews',          element: <ReviewsSection /> },
+          { path: 'global-settings',  element: <Suspense fallback={<LazyFallback />}><GeneralSection /></Suspense> },
+          { path: 'namespaces',       element: <Suspense fallback={<LazyFallback />}><NamespacesSection /></Suspense> },
+          { path: 'users',            element: <Suspense fallback={<LazyFallback />}><UsersSection /></Suspense> },
+          { path: 'oidc-providers',   element: <Suspense fallback={<LazyFallback />}><OidcProvidersSection /></Suspense> },
+          { path: 'reviews',          element: <Suspense fallback={<LazyFallback />}><ReviewsSection /></Suspense> },
         ],
       },
       { path: 'onboarding',                                              element: <ProtectedRoute><OnboardingPage /></ProtectedRoute> },
